@@ -22,24 +22,20 @@ namespace CompressFileSample
         public static void CreateZipFile(string directory, string zipFile)
         {
             InitSampleFilesForZip(directory);
-            string destDirectory = Path.GetDirectoryName(zipFile);
-            if (!Directory.Exists(destDirectory))
+            string? destDirectory = Path.GetDirectoryName(zipFile);
+            if (destDirectory != null && !Directory.Exists(destDirectory))
             {
                 Directory.CreateDirectory(destDirectory);
             }
             FileStream zipStream = File.Create(zipFile);
-            using (var archive = new ZipArchive(zipStream, ZipArchiveMode.Create))
+            using ZipArchive archive = new(zipStream, ZipArchiveMode.Create);
+            IEnumerable<string> files = Directory.EnumerateFiles(directory, "*", SearchOption.TopDirectoryOnly);
+            foreach (var file in files)
             {
-                IEnumerable<string> files = Directory.EnumerateFiles(directory, "*", SearchOption.TopDirectoryOnly);
-                foreach (var file in files)
-                {
-                    ZipArchiveEntry entry = archive.CreateEntry(Path.GetFileName(file));
-                    using (FileStream inputStream = File.OpenRead(file))
-                    using (Stream outputStream = entry.Open())
-                    {
-                        inputStream.CopyTo(outputStream);
-                    }
-                }
+                ZipArchiveEntry entry = archive.CreateEntry(Path.GetFileName(file));
+                using FileStream inputStream = File.OpenRead(file);
+                using Stream outputStream = entry.Open();
+                inputStream.CopyTo(outputStream);
             }
         }
 
@@ -62,57 +58,45 @@ namespace CompressFileSample
         public static void DecompressFile(string fileName)
         {
             FileStream inputStream = File.OpenRead(fileName);
-            using (MemoryStream outputStream = new MemoryStream())
-            using (var compressStream = new DeflateStream(inputStream, CompressionMode.Decompress))
-            {
-                compressStream.CopyTo(outputStream);
-                outputStream.Seek(0, SeekOrigin.Begin);
-                using (var reader = new StreamReader(outputStream, Encoding.UTF8, detectEncodingFromByteOrderMarks: true, bufferSize: 4096, leaveOpen: true))
-                {
-                    string result = reader.ReadToEnd();
-                    Console.WriteLine(result);
-                }
-            }
+            using MemoryStream outputStream = new MemoryStream();
+            using DeflateStream compressStream = new(inputStream, CompressionMode.Decompress);
+            compressStream.CopyTo(outputStream);
+            outputStream.Seek(0, SeekOrigin.Begin);
+            using StreamReader reader = new(outputStream, Encoding.UTF8, detectEncodingFromByteOrderMarks: true, bufferSize: 4096, leaveOpen: true);
+            string result = reader.ReadToEnd();
+            Console.WriteLine(result);
         }
 
         public static void CompressFile(string fileName, string compressedFileName)
         {
-            using (FileStream inputStream = File.OpenRead(fileName))
-            {
-                FileStream outputStream = File.OpenWrite(compressedFileName);
-                using (var compressStream = new DeflateStream(outputStream, CompressionMode.Compress))
-                {
-                    inputStream.CopyTo(compressStream);
-                }
-            }
+            using FileStream inputStream = File.OpenRead(fileName);
+
+            FileStream outputStream = File.OpenWrite(compressedFileName);
+            using DeflateStream compressStream = new(outputStream, CompressionMode.Compress);
+            inputStream.CopyTo(compressStream);
         }
 
         public static void CompressFileWithBrotli(string fileName, string compressedFileName)
         {
-            using (FileStream inputStream = File.OpenRead(fileName))
-            {
-                FileStream outputStream = File.OpenWrite(compressedFileName);
-                using (var compressStream = new BrotliStream(outputStream, CompressionMode.Compress))
-                {
-                    inputStream.CopyTo(compressStream);
-                }
-            }
+            using FileStream inputStream = File.OpenRead(fileName);
+
+            FileStream outputStream = File.OpenWrite(compressedFileName);
+            using BrotliStream compressStream = new(outputStream, CompressionMode.Compress);
+            inputStream.CopyTo(compressStream);
         }
 
         public static void DecompressFileWithBrotli(string fileName)
         {
             FileStream inputStream = File.OpenRead(fileName);
-            using (MemoryStream outputStream = new MemoryStream())
-            using (var compressStream = new BrotliStream(inputStream, CompressionMode.Decompress))
-            {
-                compressStream.CopyTo(outputStream);
-                outputStream.Seek(0, SeekOrigin.Begin);
-                using (var reader = new StreamReader(outputStream, Encoding.UTF8, detectEncodingFromByteOrderMarks: true, bufferSize: 4096, leaveOpen: true))
-                {
-                    string result = reader.ReadToEnd();
-                    Console.WriteLine(result);
-                }
-            }
+            using MemoryStream outputStream = new();
+            using BrotliStream compressStream = new(inputStream, CompressionMode.Decompress);
+
+            compressStream.CopyTo(outputStream);
+            outputStream.Seek(0, SeekOrigin.Begin);
+            using StreamReader reader = new(outputStream, Encoding.UTF8, detectEncodingFromByteOrderMarks: true, bufferSize: 4096, leaveOpen: true);
+
+            string result = reader.ReadToEnd();
+            Console.WriteLine(result);
         }
     }
 }
