@@ -24,55 +24,51 @@ namespace ReaderWriterSamples
         public static void WriteFileUsingWriter(string fileName, string[] lines)
         {
             var outputStream = File.OpenWrite(fileName);
-            using (var writer = new StreamWriter(outputStream))
-            {
-                byte[] preamble = Encoding.UTF8.GetPreamble();
-                outputStream.Write(preamble, 0, preamble.Length);
-                writer.Write(lines);
-            }
+            using var writer = new StreamWriter(outputStream);
+
+            byte[] preamble = Encoding.UTF8.GetPreamble();
+            outputStream.Write(preamble, 0, preamble.Length);
+            writer.Write(lines);
         }
 
         public static void WriteFileUsingBinaryWriter(string binFile)
         {
             var outputStream = File.Create(binFile);
-            using (var writer = new BinaryWriter(outputStream))
-            {
-                double d = 47.47;
-                int i = 42;
-                long l = 987654321;
-                string s = "sample";
+            using var writer = new BinaryWriter(outputStream);
 
-                writer.Write(d);
-                writer.Write(i);
-                writer.Write(l);
-                writer.Write(s);
-            }
+            double d = 47.47;
+            int i = 42;
+            long l = 987654321;
+            string s = "sample";
+
+            writer.Write(d);
+            writer.Write(i);
+            writer.Write(l);
+            writer.Write(s);
         }
 
         public static void ReadFileUsingBinaryReader(string binFile)
         {
             var inputStream = File.Open(binFile, FileMode.Open);
-            using (var reader = new BinaryReader(inputStream))
-            {
-                double d = reader.ReadDouble();
+            using var reader = new BinaryReader(inputStream);
 
-                int i = reader.ReadInt32();
-                long l = reader.ReadInt64();
-                string s = reader.ReadString();
-                Console.WriteLine($"d: {d}, i: {i}, l: {l}, s: {s}");
-            }
+            double d = reader.ReadDouble();
+
+            int i = reader.ReadInt32();
+            long l = reader.ReadInt64();
+            string s = reader.ReadString();
+            Console.WriteLine($"d: {d}, i: {i}, l: {l}, s: {s}");
         }
 
         public static void ReadFileUsingReader(string fileName)
         {
             var stream = new FileStream(fileName, FileMode.Open, FileAccess.Read, FileShare.Read);
-            using (var reader = new StreamReader(stream))
+            using var reader = new StreamReader(stream);
+
+            while (!reader.EndOfStream)
             {
-                while (!reader.EndOfStream)
-                {
-                    string line = reader.ReadLine();
-                    Console.WriteLine(line);
-                }
+                string? line = reader.ReadLine();
+                Console.WriteLine(line);
             }
         }
     }
